+++ momoim-core/src/main/java/com/triplem/momoim/core/domain/review/infrastructure/ReviewJpaRepository.java
@@ -1,4 +1,4 @@
-package com.triplem.momoim.core.domain.review;
+package com.triplem.momoim.core.domain.review.infrastructure;
 
 import org.springframework.data.jpa.repository.JpaRepository;
